Change double precision to float to save memory. Also prevent allocation of higher dimensional java arrays, prefering to allocate a single array in MatrixStorage and managing its access to reduce allocation overhead.

+++ jautograd-compute/src/main/java/com/ashutoshwad/utils/jautograd/ElementWiseBinaryBackwardOperation.java
@@ -69,11 +69,11 @@ public class ElementWiseBinaryBackwardOperation extends BackwardComputeOperation
                 double resultGradient = result.getGradient(row, column);
                 if (isLeftGradEnabled) {
                     double leftGradient = leftBackwardFunction.result(leftValue, rightValue, resultValue, resultGradient);
-                    left.accumulateGradient(row, column, leftGradient);
+                    left.accumulateGradient(row, column, (float) leftGradient);
                 }
                 if (isRightGradEnabled) {
                     double rightGradient = rightBackwardFunction.result(leftValue, rightValue, resultValue, resultGradient);
-                    right.accumulateGradient(row, column, rightGradient);
+                    right.accumulateGradient(row, column, (float) rightGradient);
                 }
             }
         }
